Update CS ads hold.ahk
Sometimes, weapon switching after a shot fails. This is because the "get color" command for the pixel is too slow to catch the instant un-scope from the game. However, I believe this is still better than accidentally swapping weapons while spraying with your AK due to a random black pixel. Now, there must be two black pixels at the same time for the switch to occur